requirements.txt
# Core Dependencies - Simplified and Working
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
pytz>=2023.3
